application/src/test/java/bernstein/controller/ApplicationControllerTest.java
package bernstein.controller;

import bernstein.domain.Deployment;
import bernstein.service.ApplicationService;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;

@WebMvcTest(ApplicationController.class)
@RunWith(SpringRunner.class)
public class ApplicationControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private ApplicationService applicationService;

    @Test
    public void shouldWork1() throws Exception {
        mockMvc.perform(post("/deployments/promote"));
    }

    @Test
    public void shouldWork2() throws Exception {
        mockMvc.perform(post("/deployments/deploy"));
    }
}
